+++ libs/langchain-community/src/structured_query/vectara.ts
@@ -36,7 +36,7 @@ function processValue(value: Value): string {
 }
 
 export class VectaraTranslator<
-  T extends VectaraStore
+  T extends VectaraStore,
 > extends BaseTranslator<T> {
   declare VisitOperationOutput: VectaraOperationResult;